Centre the first-run progress bar in the main screen.
Previously it was shown at the top, just below 'updating repositories', now it is
in the centre of the entire view. To make this simpler, the view was also changed
from a LinearLayout to a ConstraintLayout. Note that unlike the old RelativeLayout,
when we ask the 'app_list' to be below the 'banner_updating_repos', it works as
expected when the banner is not shown. That is to say, it will correctly anchor the
top of the app list to the top parent component if the banner is not shown, but will
sit it below the banner if the banner is shown.