Retry downloads from the original repo address if the file was not found on a mirror
After a repository was updated, the mirrors need some time to sync the changes. In the meantime, users who got the index file from the main repo can try to download an APK from one of the outdated mirrors.
The app proceeds by showing a toast saying "The requested file was not found". That is bad UX, because the users are notified notified about available updates which they want to install, but F-Droid is unable to do that.

This issue is resolved by retrying the download using the original repo address.