block adding repos when Unknown Sources is disable by Device Admin
Device Admin apps aka Mobile Device Management can set all sorts of
restrictions on the device.  One of these is to globally disable the use
of Unknown Sources.  The other is to restrict a user on the device from
using Unknown Sources.  Both of these are supported.  When either or both
are set, then F-Droid will not allow adding new repos, either via Intent or
via the UI.

refs #2238
https://gitlab.com/groups/CalyxOS/-/epics/25
https://gitlab.com/CalyxOS/calyxos/-/issues/611