nearby: ensure widgets are not null before trying to configure them
There are lots of async events in this system, and the user can also change
the views while things are working.  The ViewGroup `container` can have all
its widgets removed and replaced by a new view at any point.  Therefore,
any widget config that is based on fetching it from `container` must check
that the result is not null before trying to config it.