[db] Add IndexV2Updater and RepoUpdater
The IndexV2Updater fetches the new entry.jar, verifies it and then either processes a diff or the full index.

The RepoUpdater is a convenience class that encapsulates updating repos with potentially more than one format. It tries v2 first (in case the repo has upgraded) and falls back to v1 if the repo isn't known to support v2 already (downgrade protection).