include only selected IPFS Gateways #2504
I reviewed a bunch of gateways in
https://gitlab.com/fdroid/fdroidclient/-/issues/2504#note_1289737638

And these look good.  They match the standard of when we first started
adding official mirrors, but I was not able to confirm their policies are
as good as the official mirrors.  This stuff is still opt-in and beta, so
this seems like a good place to start the process.