[app] Allow background downloads/installs on Android 8
It is not allowed anymore to start services while the app is in the background. To work around this, the InstallManagerService was turned into a singleton (it needs to receive broadcasts for as long as possible) and DownloaderService was turned into a JobIntentService which we are still allowed to "start" from the background. This start might be delayed, but better late than never.

This is a temporary workaround that allows us to start this service from the background. The long term plan is to use WorkManager for it as well.