🥙 added UI for configuring custom IPFS gateways
This change adds an additional activity for adding custom IPFS gateways.
Also the IPFS Gateway settings list custom gateways now and offen an
action to delete them. The UI is mostly finished now. It's storing its
state in `Preferences`, but the stored values are not used by the
downloader yet.