Prevent users from accidentally creating a category search filter
Previously, users could accidentally trigger their search input to be interpreted as category when they used a colon in their search term. The colon was used as separator between search filter and search query, thus the text before the colon was interpreted as category. The colon is replaced by the non-printable NUL character (char value 0). A constant for the new filter separator is introduced  and used to increase maintainability.